[move-ide] Fixed a problem related to handling package identity
## Description 

This PR fixes a problem with move-analyzer crashing if a dependency
defined one package name in the source but the main package overrode
this name with a new one. The gist of the problem was that we used
package name to represent its identity which is not reliable due to
above-mentioned renamings. What we need to do instead is to rely on
package addresses whenever they are available.

## Test plan 

Added a new test that was crashing the analyzer before this PR's fix.