[Doc] rewrite consensus section
## Description 

I decided to remove most the previous consensus page, because they seem
to be confusing.
Only a short overview section is added to replace the previous page. We
can include more images later on. Not sure if we want very detailed docs
here though, which can be hard to keep updated.

## Test plan 

n/a

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------