[CI] Add external PRs to the github-issues workflow
## Description 

When a PR is created, this workflow will pick it up and find if it is
from someone in the HaneulLabs organization or not. If it is not, then
it will push a Slack message to notify us of this PR coming from an
external contributor.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: