Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/zksend@0.6.0

### Minor Changes

-   aedbb10a40: Add `createSendToAddressTransaction` method

### Patch Changes

- 666a674a56: Fix assets extracted from transactionBlocks using type
instead of objectType from object changes"

## haneullabs/create-dapp@0.2.14

### Patch Changes

-   haneullabs/dapp-kit@0.12.12

## haneullabs/dapp-kit@0.12.12

### Patch Changes

-   Updated dependencies [666a674a56]
-   Updated dependencies [aedbb10a40]
    -   haneullabs/zksend@0.6.0