[consensus] Add ReputationScoreCalculator & ScoringStrategy
## Description 
This is one of many PRs to implement Leader Scoring & Leader Schedule
feature for Mysticeti.
- [Design
Doc](https://www.notion.so/haneullabs/Leader-Schedule-Scoring-538c65cb370b4644944a94485efc4979?pvs=4)
- [Prototype PR](https://github.com/GeunhwaJeong/haneul/pull/16975)

Notes:
- Unscored subdags will eventually be cached in `DagState` and when
there are enough commits to trigger a leader schedule change the
`ReputationScoreCalculator` will be used to calculate the
`ReputationScores`. As can be seen in the flow diagram below.
- Added multiple scoring strategies so that we can experiment with them
quickly until we decided on the strategy that is best to use. By default
we will use the `VoteScoringStrategy` when everything is hooked up as it
had the best early
[results](https://metrics.haneul.io/goto/yeUlNPxIR?orgId=1).
- We are sharing the `UniversalCommitter` inside of the
`ReputationScoreCalculator` so we have consistency with the leaders that
we elect when scoring.

![Screenshot 2024-04-03 at 7 59
09 PM](https://github.com/GeunhwaJeong/haneul/assets/97870774/b91931ab-7f92-49e8-90ea-2d0e8476d1d8)

## Test Plan 

unit tests