Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/dapp-kit@0.13.0

### Minor Changes

- 251d8be42a: The `zksend` property on the `WalletProvider` has been
replaced with a `stashedWallet` option.

### Patch Changes

-   Updated dependencies [251d8be42a]
    -   haneullabs/zksend@0.8.0

## haneullabs/zksend@0.8.0

### Minor Changes

-   251d8be42a: Update zkSend Wallet references to Stashed.

## haneullabs/create-dapp@0.2.18

### Patch Changes

-   Updated dependencies [251d8be42a]
    -   haneullabs/dapp-kit@0.13.0