[haneul-framework] Reserve addresses for the deny list
## Description 

Added a list of well-known package and object addresses, under the
constant `RESERVED`. It would be meaningless to add them to the deny
list; as such, they are reserved for internal use--specifically as
sentinel values.

## Test plan 

- Added tests 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------