Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/create-dapp@0.2.19

### Patch Changes

-   Updated dependencies [e52c7068a1]
    -   haneullabs/dapp-kit@0.13.1

## haneullabs/dapp-kit@0.13.1

### Patch Changes

-   e52c7068a1: Update stashed logo and fix detection of stashed wallet
-   Updated dependencies [e52c7068a1]
    -   haneullabs/zksend@0.8.1

## haneullabs/zksend@0.8.1

### Patch Changes

-   e52c7068a1: Update stashed logo and fix detection of stashed wallet