Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/create-dapp@0.2.20

### Patch Changes

-   Updated dependencies [d838743288]
    -   haneullabs/dapp-kit@0.13.2

## haneullabs/dapp-kit@0.13.2

### Patch Changes

- d838743288: Fix broken theme style tag in canary versions of React
when the provider is placed outside of the body tag
-   Updated dependencies [be9ddf1767]
    -   haneullabs/zksend@0.8.2

## haneullabs/zksend@0.8.2

### Patch Changes

-   be9ddf1767: Update stashed icon