[docs][easy] Removing Pattern topics
## Description 
The pattern topics are covered in the Haneul Book so the decision was made
to remove. Initial work only removed the navigation listing, not the
actual topics. This PR also adds a general redirect. The pages aren't
1:1 and there's a risk the URLs change anyway, so using the main page as
the destination. Should be low impact regardless.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: