[Consensus 2.0] Calculate committed sub dag digest
## Description 

Currently the default digest is being used on the mysticeti committed
subdag making the fork detection in checkpoints weaker as included in
the `ConsensusCommitInfo`. This PR is calculating the committed sub dag
digest and enabling it behind a feature flag. It has to be noted that
the reputation scores are not included in the sub dag digest as during
crash recovery we don't restore the actual scores for every committed
sub dag, but only restoring the last ones and sent as part of the last
committed sub dag.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: