[move-ide] Embed move-analyzer within Haneul CLI
## Description 

The `move-analyzer` binary was recently moved to `crates` to support its
versioning according to the same scheme as the `haneul` binary. The goal is
to distribute `move-analyzer` in two ways:
- embed it with VSCode extension for those that just want to use the IDE
and don't want or (yet) need Haneul CLI
- via Haneul CLI (so that each time CLI is updated, `move-analyzer` binary
is updated as well)

Out of these two options, the newer one (according to the version
number) will be chosen to be used by the VSCode extension.

This PR embeds `move-analyzer` within Haneul CLI so that it can be
distributed via brew and chocolatey. The difference in binary size is
rather small 110MB vs 108MB on a Mac