[move][ide] Report missing match arms as IDE annotations
## Description 

This reports missing match arms as IDE annotations, so the IDE can
suggest adding them. It only does the outer layer, a bit like Rust.

Note that if an inner pattern is not exhaustive, it will simply report
that as an error, no suggestion (on parity with rust analyzer).

## Test plan 

This is currently not tested. I plan to put up another diff to allow us
to test all IDE annotations by reporting them as diagnostics when in
test mode. When I do that, I will also write tests for this diff.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: