[Metrics] add bounded and unbounded mpsc channel wrappers
## Description 

Majority of the implementation is adapted from the existing metered
bounded channel. The main advantage of the new implementation is that it
eliminates the need to create separate metrics for each channel.

Channels in `consensus-core` are migrated to the new monitored channel
wrappers. This adds monitoring for unbounded channels.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: