Bridge test env
## Description 

Increase test coverage and create a test environment for the bridge that
should make it easier to test.
At this point once we have API to sign messages we can reach 100%
coverage. All missing tests are related to signature generation,
function depending on that, or versioning. API to sign messages and
create pub/priv keys are coming that would allow us to create a great
testing environment and experience.
Current coverage is as follows
```
+-------------------------+
| Move Coverage Summary   |
+-------------------------+
Module 000000000000000000000000000000000000000000000000000000000000000b::treasury
>>> % Module coverage: 100.00
Module 000000000000000000000000000000000000000000000000000000000000000b::message_types
>>> % Module coverage: 100.00
Module 000000000000000000000000000000000000000000000000000000000000000b::chain_ids
>>> % Module coverage: 100.00
Module 000000000000000000000000000000000000000000000000000000000000000b::message
>>> % Module coverage: 100.00
Module 000000000000000000000000000000000000000000000000000000000000000b::limiter
>>> % Module coverage: 100.00
Module 000000000000000000000000000000000000000000000000000000000000000b::crypto
>>> % Module coverage: 100.00
Module 000000000000000000000000000000000000000000000000000000000000000b::committee
>>> % Module coverage: 100.00
Module 000000000000000000000000000000000000000000000000000000000000000b::bridge
>>> % Module coverage: 51.21
+-------------------------+
| % Move Coverage: 82.32  |
+-------------------------+
```

The test environment needs more work but it will come in time. For now
this work was mostly to reach a good test coverage.
It's important to check this in for 2 main reasons: it increases test
coverage significantly and allows others to work on it and improve it.

## Test plan 

This is it

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: