[CLI] fix command default value problem
## Description 

[CLI] fix `haneul keytool zk-login-sign-and-execute-tx` command, should set
the default value of `--test-multisig` to `false`, because it cannot be
set to `false` in the command line. So current the generated zkLogin
addresses will always be multisig address, and it is not possible to be
single-sig address anymore.

The `--fixed` parameter also has the same problem.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: `haneul keytool zk-login-sign-and-execute-tx`: set the default
value for `--test-multisig` and `--fixed` flags to `false`, because it
cannot be set to false in the command line.
- [ ] Rust SDK:

---------