improve coin and balance query performance
## Description 

We need a dedicated index for `coin_type`. This can be a partial index
`WHERE coin_type IS NOT NULL`

## Test plan 

Manually ran benchmark suite and verified that all benchmarked queries
did not timeout and completed within a sensible amount of time.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: