[move] Use clever error locations for macro-rewritten lines in Move unit test errors
## Description 

This changes the way we report source locations for clever errors coming
from an `assert!(<bool>)` within a macro. In particular, since the line
number is rewritten during macro-expansion, we use this as the location
of the error in unit tests when reporting the error if there is no
overlap between the line number from the clever assertion error and the
"actual" error.

Note that it was easiest to do this by plumbing in the `MappedFiles`
instead of using the `FilesSourcesText` in unit tests anymore. This
simplified a number of other places as well (e.g., for getting the line
number from a location).

## Test plan 

Made sure existing tests continued to pass, and added tests for the new
behavior.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: