[move-ide] Added support for intelligent dot auto-completion
## Description 

This PR stands on the shoulders of previously done compiler-side work
and simply adds the IDE-side of things to support intelligent
auto-completion.

Works for dot completion without valid identifier following the dot:


![image](https://github.com/GeunhwaJeong/haneul/assets/1724397/5de46118-e79a-4bb8-aa88-8204d7d5d0b9)

It also works for dot completion when the following identifier is
already (partially) typed:


![image](https://github.com/GeunhwaJeong/haneul/assets/1724397/6d3e11c7-5523-4171-9671-4f4da14703ec)

Note that additional information (in addition to just name to be
completed is available): target function (for methods) as well as the
target function signature/type (for methods/fields)

Field completions are inserted "as is" (just the field name) but method
completions are inserted as snippets listing parameter names for a given
method. The following screenshot shows the effect of accepting the
auto-completion suggestion from the screenshot directly above (`param`
has been inserted automatically and did not have to be typed by the
programmer; if more params were available, their names would be inserted
as well and the programmer can move to the next one by pressing TAB):


![image](https://github.com/GeunhwaJeong/haneul/assets/1724397/8232cd6e-7593-49d2-a98d-58b217dbd241)

This PR also adds missing auto-completion info in the compiler (method
names in addition to target function names) as well additional info (a
struct containing fields provided to auto-completion query)

## Test plan 

Correctness of auto-completion items being provided to the IDE is
already tested on the compiler side, and a new test there was added.

Still, added an IDE test for things that are IDE-specific that cannot be
tested on the compiler-side. We are considering building a more powerful
testing framework for IDE unit tests, but for not (considering
compiler-side testing), this added test will hopefully suffice.